refactor: Extract cache module and use reth-primitives-traits v1 for legacy migration
- Move cache format handling (load, save, versioned formats) from lib.rs
  into its own cache.rs module
- Use reth-primitives-traits v1.11.3 (as reth-primitives-traits-v1) for
  deserializing the serde_bincode_compat block format in v1/v2 cache files
- Remove alloy-consensus direct dependency (no longer needed)
