+++ crates/host/Cargo.toml
@@ -11,7 +11,6 @@ unsafe-pre-merge = ["guests/unsafe-pre-merge", "zeth-core/unsafe-pre-merge"]
 
 [dependencies]
 alloy = { workspace = true, features = ["provider-debug-api"] }
-alloy-consensus = { workspace = true, features = ["serde", "serde-bincode-compat"] }
 anyhow = { workspace = true }
 blst = { workspace = true }
 bytemuck = { workspace = true }
@@ -19,6 +18,7 @@ clap = { workspace = true, features = ["derive", "env"] }
 guests = { workspace = true }
 reth-chainspec = { workspace = true }
 reth-ethereum-primitives = { workspace = true }
+reth-primitives-traits-v1 = { package = "reth-primitives-traits", git = "https://github.com/paradigmxyz/reth.git", tag = "v1.11.3", features = ["serde-bincode-compat"] }
 risc0-zkvm = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
